bump to v0.5.0: _xp_seq INT->BIGINT, 11 bug fixes, SQL validation
- _xp_seq column type changed from INT (2.1B limit) to BIGINT
- All C source files migrated: int32->int64 for seq values
- group_stats.max_seq upgraded to BIGINT
- 11 bug fixes from production audit (H1-H7, L5-L6, C1-C2, M1-M3)
- E13/E17: order_by type validation in xpatch.configure()
- New SQL migration: pg_xpatch--0.4.0--0.5.0.sql
- Reverted 0.4.0 SQL files to original (no retroactive edits)
- 419 tests passing, 0 failures

+++ src/xpatch_stats_cache.h
@@ -31,7 +31,7 @@ void xpatch_stats_cache_update_group(
     Oid relid,
     XPatchGroupHash group_hash,
     bool is_keyframe,
-    int32 max_seq,
+    int64 max_seq,
     int64 raw_size,
     int64 compressed_size,
     double avg_delta_tag
@@ -52,7 +52,7 @@ void xpatch_stats_cache_delete_table(Oid relid);
  * Get max_seq for a group from stats cache.
  * Returns -1 if not found.
  */
-int32 xpatch_stats_cache_get_max_seq(Oid relid, XPatchGroupHash group_hash);
+int64 xpatch_stats_cache_get_max_seq(Oid relid, XPatchGroupHash group_hash);
 
 /*
  * Check if stats exist for a table.
